+++ backend/src/main/java/reviewme/auth/controller/AuthController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import reviewme.auth.domain.GitHubMember;
 import reviewme.auth.service.AuthService;
-import reviewme.global.session.SessionManager;
+import reviewme.security.session.SessionManager;
 import reviewme.reviewgroup.service.dto.CheckValidAccessRequest;
 
 @RestController
