+++ backend/src/main/java/reviewme/config/WebConfig.java
@@ -5,9 +5,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import reviewme.auth.controller.GuestReviewGroupSessionResolver;
-import reviewme.auth.controller.LoginMemberSessionResolver;
-import reviewme.global.session.SessionManager;
+import reviewme.security.resolver.GuestReviewGroupSessionResolver;
+import reviewme.security.resolver.LoginMemberSessionResolver;
+import reviewme.security.session.SessionManager;
 
 @Configuration
 @RequiredArgsConstructor
