+++ backend/src/main/java/reviewme/review/controller/ReviewController.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import reviewme.auth.controller.GuestReviewGroupSession;
-import reviewme.auth.controller.LoginMemberSession;
-import reviewme.auth.controller.dto.GuestReviewGroup;
-import reviewme.auth.controller.dto.LoginMember;
+import reviewme.security.resolver.GuestReviewGroupSession;
+import reviewme.security.resolver.LoginMemberSession;
+import reviewme.security.resolver.dto.GuestReviewGroup;
+import reviewme.security.resolver.dto.LoginMember;
 import reviewme.review.service.ReviewDetailLookupService;
 import reviewme.review.service.ReviewGatheredLookupService;
 import reviewme.review.service.ReviewListLookupService;
