+++ backend/src/main/java/reviewme/reviewgroup/controller/ReviewGroupController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import reviewme.auth.controller.LoginMemberSession;
-import reviewme.auth.controller.dto.LoginMember;
+import reviewme.security.resolver.LoginMemberSession;
+import reviewme.security.resolver.dto.LoginMember;
 import reviewme.reviewgroup.service.ReviewGroupLookupService;
 import reviewme.reviewgroup.service.ReviewGroupService;
 import reviewme.reviewgroup.service.dto.ReviewGroupCreationRequest;
