+++ backend/src/main/java/reviewme/security/resolver/GuestReviewGroupSession.java
@@ -1,4 +1,4 @@
-package reviewme.auth.controller;
+package reviewme.security.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
